feat(stats): make common functionality to return the value of any stat taking into account item modifiers

+++ rust/src/stats/modifier/custom/implementation.rs
@@ -1,11 +1,12 @@
 use crate::stats::modifier::custom::definition::CustomModifier;
 
-pub(crate) fn get_implementation_for_custom_modifier(custom_mod: CustomModifier) -> fn(u16) -> i16 {
+pub(crate) fn get_implementation_for_custom_modifier(custom_mod: CustomModifier) -> fn(i16) -> i16 {
     match custom_mod {
         CustomModifier::ONE => one_modifier,
     }
 }
 
-fn one_modifier(stat: u16) -> i16 {
-    (stat as i16 * -1) + 1
+// Sets the stat value to 1 no matter the original value
+fn one_modifier(stat: i16) -> i16 {
+    (stat * -1) + 1
 }
